requirements.txt
Flask
werkzeug
holidays
pypdf
reportlab
argon2-cffi